refactor(cli): move markdown helpers into utils
Move markdown utilities and tests from src/cli/markdown to src/cli/utils.

Update cli exports and skills imports to use the new module path.
