+++ src/cli.ts
@@ -5,7 +5,7 @@ export * from './cli/eval/eval.constants.ts'
 export * from './cli/eval/eval.schemas.ts'
 export * from './cli/eval/eval.ts'
 export * from './cli/eval/eval.utils.ts'
-export * from './cli/markdown/markdown.ts'
 export * from './cli/skills/skills.schema.ts'
 export * from './cli/skills/skills.ts'
 export * from './cli/utils/cli.ts'
+export * from './cli/utils/markdown.ts'
